move `IMainProcessService`
