+++ src/vs/code/node/sharedProcess/contrib/storageDataCleaner.ts
@@ -17,7 +17,7 @@ import { INativeHostService } from 'vs/platform/native/common/native';
 
 /* eslint-disable local/code-layering, local/code-import-patterns */
 // TODO@bpasero layer is not allowed in utility process//
-import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/services';
+import { IMainProcessService } from 'vs/platform/ipc/common/mainProcessService';
 
 export class UnusedWorkspaceStorageDataCleaner extends Disposable {
