+++ src/vs/platform/native/electron-sandbox/nativeHostService.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
-import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/services';
+import { IMainProcessService } from 'vs/platform/ipc/common/mainProcessService';
 import { INativeHostService } from 'vs/platform/native/common/native';
 
 // @ts-ignore: interface is implemented via proxy
