+++ src/vs/workbench/contrib/webview/electron-sandbox/windowIgnoreMenuShortcutsManager.ts
@@ -6,7 +6,7 @@
 import { isMacintosh } from 'vs/base/common/platform';
 import { ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/services';
+import { IMainProcessService } from 'vs/platform/ipc/common/mainProcessService';
 import { INativeHostService } from 'vs/platform/native/common/native';
 import { IWebviewManagerService } from 'vs/platform/webview/common/webviewManagerService';
